Conflicts:
	fst/CMakeLists.txt
	fst/XrdFstOfsFile.cc
	fst/XrdFstOfsFile.hh
	fst/XrdFstOssFile.cc
	fst/io/FileIo.hh
	fst/layout/RaidMetaLayout.cc
	fst/layout/ReedSLayout.cc
	fst/layout/ReplicaParLayout.cc
	mgm/XrdMgmOfs.cc
	mgm/proc/user/File.cc

+++ fst/layout/Layout.cc
@@ -41,9 +41,8 @@ mOfsFile (file)
 //------------------------------------------------------------------------------
 // Constructor
 //------------------------------------------------------------------------------
-
 Layout::Layout (XrdFstOfsFile* file,
-                int lid,
+                unsigned long lid,
                 const XrdSecEntity* client,
                 XrdOucErrInfo* outError,
                 eos::common::LayoutId::eIoType io,
@@ -65,7 +64,6 @@ mTimeout(timeout)
 //------------------------------------------------------------------------------
 // Destructor
 //------------------------------------------------------------------------------
-
 Layout::~Layout () {
   // empty
 }
